# backend/openapi_v3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "backend",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nresolvet customer support ticket backend"
  },
  "paths": {
...
    "/api/v1/tickets/{id}": {
      "get": {
        "summary": "/api/v1/tickets/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": []
      },
      "put": {
        "summary": "/api/v1/tickets/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": []
      }
    },
...
  },
...
}